% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_district_to_points.R
\name{add_district_to_points}
\alias{add_district_to_points}
\title{Attach district names to destination points}
\usage{
add_district_to_points(population, districts, crs_out = 4326)
}
\arguments{
\item{population}{Tibble of destination points (must include lon, lat)}

\item{districts}{An sf object of Malawi districts}

\item{crs_out}{Output CRS for coordinates (default: 4326)}
}
\value{
A tibble with district column added
}
\description{
Attach district names to destination points
}
