% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors-airtable.R
\name{add_editor_airtable_data}
\alias{add_editor_airtable_data}
\title{Add additional columns to 'editors' data from rOpenSci's airtable database,
and remove any "emeritus" editors.}
\usage{
add_editor_airtable_data(editors, airtable_id)
}
\arguments{
\item{editors}{The `data.frame` of editors returned as the "status" component
from \link{editor_status}.}
}
\value{
A modified version of `editors` with additional columns.
}
\description{
Add additional columns to 'editors' data from rOpenSci's airtable database,
and remove any "emeritus" editors.
}
