% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_github}
\alias{add_github}
\title{Add GitHub username or organization name to packages}
\usage{
add_github(repo, org = "ropensci")
}
\arguments{
\item{repo}{Name of the repository (without org name)}

\item{org}{= "ropensci" Your GitHub org name or username. Use this function only if all the repos are part of the same account. Otherwise add manually.}
}
\description{
Add GitHub username or organization name to packages
}
\examples{
\dontrun{
add_github("alm", "ropensci")
}
}

