% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_r5r_network.R
\name{build_r5r_network_cached}
\alias{build_r5r_network_cached}
\title{Build or load an r5r transport network}
\usage{
build_r5r_network_cached(data_path)
}
\arguments{
\item{data_path}{Path to the folder containing OSM and GTFS data}
}
\value{
An r5r network object
}
\description{
Build or load an r5r transport network
}
