% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community-leaderboard.R
\name{community_data}
\alias{community_data}
\title{Data for community leaderboard}
\usage{
community_data(airtable_id, quiet = FALSE)
}
\arguments{
\item{airtable_id}{ID of Airtable base}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
A named list with one element for each community member and integer
values for total number of packages submitted; total number of reviews
conducted; and an additional logical flag indicating whether or not that
person is or was a member of the editorial team. These data are returned as
a raw list, because they are processed directly in Javascript.
}
\description{
This is only slightly modified from \link{review_history}, through adding
additional Airtable data to identify current editorial team.
}
