% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_subgroups.R
\name{create_population_subgroup_rasters}
\alias{create_population_subgroup_rasters}
\title{Create and write population subgroup rasters from WorldPop files}
\usage{
create_population_subgroup_rasters(
  country,
  raw_dir,
  filepath,
  year,
  out_dir = here::here("data", "population")
)
}
\arguments{
\item{country}{Country name used to derive ISO3 code.}

\item{raw_dir}{Directory containing the unzipped WorldPop .tif files}

\item{filepath}{The filepath of the ZIP file downloaded from WorldPop.}

\item{year}{WorldPop data year.}

\item{out_dir}{Directory to write output rasters (default: "data")}
}
\value{
A combined stack raster for each population sub group
}
\description{
This function reads the unzipped WorldPop .tif files for a specified country
and creates combined stack rasters for different population subgroups
}
