% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-package.R
\docType{package}
\name{dashboard-package}
\alias{dashboard-package}
\title{dashboard: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-review-tools/dashboard}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
