% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard-fn.R
\name{dashboard}
\alias{dashboard}
\title{Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.}
\usage{
dashboard(open_only = TRUE, browse = TRUE)
}
\arguments{
\item{open_only}{If `TRUE` (default), only extract information for currently
open issues.}

\item{browse}{If `TRUE` (default), open the results as a \pkg{DT} `datatable`
HTML page in default browser.}
}
\value{
A `data.frame` with one row per issue and some key statistics.
}
\description{
Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.
}
