% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dashboard_launch}
\alias{dashboard_launch}
\title{dashboard_launch generates the dashboard and launchs the local server}
\usage{
dashboard_launch(browse = TRUE)
}
\arguments{
\item{browse}{boolean for launching the local server}
}
\description{
\code{dashboard_launch} writes all files for the web page and launchs the local server
}
\examples{
dashboard_open(data=iris) # other options: pathoutput=getwd() ...
dcpiechart(x=names(iris)[5])
dcbarchart(x=names(iris)[1] , gap=75)
dcpiechart(x=names(iris)[2])
dctable(index=names(iris)[5])
dashboard_launch(browse = FALSE) # Just generates files. Server is not launched
}

