% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dashboard_open}
\alias{dashboard_open}
\title{dashboard_open innitializes a new dashboard}
\usage{
dashboard_open(data, title = "Dashbord test", filename = "dashboard",
  pathoutput = tempdir(), outerwidth = 960, outerheight = 700)
}
\arguments{
\item{data}{data frame used for drawing a dashboard}

\item{title}{character for the title of the generated dashboard}

\item{filename}{character for the name of the generated html file}

\item{pathoutput}{character for the output path of generated files}

\item{outerwidth}{integer for the outer width (in pixel) of the web page}

\item{outerheight}{integer for the outer height (in pixel) of the web page}
}
\description{
\code{dashboard_open} sets the configuration for the web page
}
\examples{
dashboard_open(data=iris) # other options: pathoutput=getwd() ...
dcpiechart(x=names(iris)[5])
dcbarchart(x=names(iris)[1] , gap=75)
dcpiechart(x=names(iris)[2])
dctable(index=names(iris)[5])
dashboard_launch(browse = FALSE) # Just generates files. Server is not launched
}

