% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_databaseConnect.R
\name{databaseConnect}
\alias{databaseConnect}
\title{databaseConnect}
\usage{
databaseConnect(sensorCodes)
}
\value{
The return value, if any, from executing the function.
}
\description{
Establishes connection to the scevo PostgreSQL database and fetches data for one or more sensor codes
}
