% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dctable}
\alias{dctable}
\title{dctable add a data table to the dashboard}
\usage{
dctable(index = listcol[1], listcol = dashboard.env$listcol,
  title = "data list table", spansize = 12)
}
\arguments{
\item{index}{column name of data frame \code{data} for indexing the data table}

\item{listcol}{vector of column names of data frame \code{data} to display in the data table}

\item{title}{character for the title of the generated data table}

\item{spansize}{integer between 1 to 12 for the width of the element in the web page}
}
\description{
\code{dctable} displays a data table
}
\examples{
dashboard_open(data=iris) # other options: pathoutput=getwd() ...
dcpiechart(x=names(iris)[5])
dcbarchart(x=names(iris)[1] , gap=75)
dcpiechart(x=names(iris)[2])
dctable(index=names(iris)[5])
dashboard_launch(browse = FALSE) # Just generates files. Server is not launched
}

