% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_pbf.R
\name{download_geofabrik_pbf}
\alias{download_geofabrik_pbf}
\title{Download Geofabrik PBF for a country}
\usage{
download_geofabrik_pbf(country, dest_dir, overwrite = FALSE, verify_md5 = TRUE)
}
\arguments{
\item{country}{Country name (e.g., "Nepal")}

\item{dest_dir}{Destination directory for the PBF file}

\item{overwrite}{If FALSE, skip download when file already exists}

\item{verify_md5}{If TRUE, validate downloaded file against Geofabrik MD5}
}
\value{
Path to local .osm.pbf file
}
\description{
Download Geofabrik PBF for a country
}
