% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_subgroups.R
\name{download_worldpop}
\alias{download_worldpop}
\title{Download and extract population subgroup files from a specified URL.}
\usage{
download_worldpop(
  iso_code,
  year,
  release,
  version = "v1",
  dest_dir = here::here("data", "raw")
)
}
\arguments{
\item{year}{WorldPop data year.}

\item{dest_dir}{The directory to save the downloaded ZIP file (default: "data
/raw").}

\item{country}{Country name used to derive ISO3 code when filepath is missing.}

\item{filepath}{The filepath of the ZIP file downloaded from WorldPop.}

\item{zip_name}{The name to save the downloaded ZIP file as (default: basename
of the URL).}

\item{unzip_folder}{The directory to extract the ZIP file to (default: same as
dest_dir).}
}
\value{
The path to the folder where the files were extracted, or NULL if there
was an error.
}
\description{
This function checks if the target folder for the extracted files already exists.
If it does, it skips the download and extraction steps. Otherwise, it downloads the
ZIP file from the provided URL, extracts its contents to a specified directory,
and then removes the ZIP file.
}
