% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{ed_rev_durations}
\alias{ed_rev_durations}
\title{Extract average duration of reviews for each editor}
\usage{
ed_rev_durations(nyears = 2)
}
\arguments{
\item{nyears}{Length of time prior to current date over which average
durations should be calculated.}
}
\value{
A `data.frame` with two columns of 'editor' and 'duration' in days.
}
\description{
Note that is extracts data for all editors, including those who may not be
part of current editorial team. Current team members can be identified from
data returned from the \link{editor_status} function.
}
