% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{editor_status}
\alias{editor_status}
\title{Generate a summary report of current state of all rOpenSci editors}
\usage{
editor_status()
}
\value{
A `data.frame` with one row per issue and some key statistics.
}
\description{
Generate a summary report of current state of all rOpenSci editors
}
