% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors-airtable.R
\name{editor_vacation_status}
\alias{editor_vacation_status}
\title{Get current vacation status of all rOpenSci editors.}
\usage{
editor_vacation_status(airtable_id)
}
\arguments{
\item{airtable_id}{The 'id' string of the airtable table.}
}
\value{
A `data.frame` with one row per editor and information on current
vacation status.
}
\description{
Data are obtained both from airtable and current slack status. If either of
these indicate that an editor is away, on vacation, or otherwise unavailable,
then the `away` column will be set to `TRUE`.
}
