% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_server.R
\name{filter_ttm_data}
\alias{filter_ttm_data}
\title{Filter TTM data.table by district and facility type}
\usage{
filter_ttm_data(ttm, selected_district, facility_types, all_district_value)
}
\arguments{
\item{ttm}{A data.table with columns \code{district} and \code{facility_type}}

\item{selected_district}{The currently selected district value from the UI}

\item{facility_types}{Character vector of selected facility types}

\item{all_district_value}{Sentinel string representing "all areas"}
}
\value{
A filtered data.table
}
\description{
Filter TTM data.table by district and facility type
}
