% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_district_to_points.R
\name{find_crs}
\alias{find_crs}
\title{Find the appropriate CRS for a set of districts}
\usage{
find_crs(districts)
}
\arguments{
\item{districts}{An sf object of districts}
}
\value{
A character string with the EPSG code of the appropriate CRS
}
\description{
Find the appropriate CRS for a set of districts
}
