% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{format_list_and}
\alias{format_list_and}
\title{Join a vector of items into a human-readable "a, b and c" list}
\usage{
format_list_and(x)
}
\description{
Join a vector of items into a human-readable "a, b and c" list
}
