% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_html}
\alias{generate_html}
\title{Generates a static html dashboard from GitHub stats and CRAN downloads}
\usage{
generate_html(out, path = "/tmp", browse = TRUE)
}
\arguments{
\item{out}{A list object generated by github_stats()}

\item{path}{Folder where you need the dashboard rendered}

\item{browse}{= TRUE Automatically open index.html in the default browser. Set to \code{FALSE} to disable.}
}
\description{
Generates a static html dashboard from GitHub stats and CRAN downloads
}
\examples{
\dontrun{
generate_html(results)
}
}

