% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_districts.R
\name{get_districts}
\alias{get_districts}
\title{Load districts from OpenStreetMap or local file}
\usage{
get_districts(country, filepath, name_col, admin_level = 4, crs = 4326)
}
\arguments{
\item{country}{The name of the country to receive boundaries for}

\item{filepath}{Path to the boundary file}

\item{name_col}{Name of the column in the file to use as the district label}

\item{admin_level}{OSM admin level (default 4 for districts)}

\item{crs}{Coordinate Reference System to return (default: 4326)}
}
\value{
An sf object of selected boundaries
}
\description{
Load districts from OpenStreetMap or local file
}
