% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_districts.R
\name{get_file_districts}
\alias{get_file_districts}
\title{Load district boundaries from a local file (shp, gpkg, geojson)}
\usage{
get_file_districts(filepath, name_col, crs = 4326)
}
\arguments{
\item{filepath}{Path to the boundary file}

\item{name_col}{Name of the column in the file to use as the district label}

\item{crs}{Coordinate Reference System to return (default: 4326)}
}
\value{
An sf object of selected boundaries
}
\description{
Load district boundaries from a local file (shp, gpkg, geojson)
}
