% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_pbf.R
\name{get_geofabrik_index}
\alias{get_geofabrik_index}
\title{Fetch Geofabrik index as a data frame}
\usage{
get_geofabrik_index(index_url = "https://download.geofabrik.de/index-v1.json")
}
\arguments{
\item{index_url}{URL to Geofabrik GeoJSON index}
}
\value{
A data frame with Geofabrik extracts metadata
}
\description{
Fetch Geofabrik index as a data frame
}
