% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors-slack.R
\name{get_slack_editors_status}
\alias{get_slack_editors_status}
\title{Get current slack status of all editors}
\usage{
get_slack_editors_status()
}
\value{
A `data.frame` with one row row each editor, and columns of:
\itemize{
   \item `id` The Slack ID of the editor
   \item `name` The name of the editor
   \item `real_name` The real name of the editor
   \item `status` The current status of the editor
}
}
\description{
The API endpoints for slack are described at \url{https://api.slack.com/web}.
All API endpoints are in the form
`https://slack.com/api/METHOD_FAMILY.method`.
}
\details{
This function requires an environment variable, "SLACK_TOKEN", to
authenticate; see \url{https://api.slack.com/docs/token-types}. The current
token is a "bot token" for the "status-app" app,
\url{https://api.slack.com/apps/A06TMRCU40J/oauth}.
}
