% Generated by roxygen2 (4.0.2): do not edit by hand
\name{github_auth}
\alias{github_auth}
\title{Authenticate with GitHub and retrieve a token}
\usage{
github_auth(gh_appname = getOption("gh_appname"),
  gh_id = getOption("gh_id"), gh_secret = getOption("gh_secret"))
}
\arguments{
\item{gh_appname}{Github app name}

\item{gh_id}{GitHub ID}

\item{gh_secret}{GitHub secret}
}
\description{
Create a new application in your github settings and add the app name, id, and secret to your \code{.rprofile} as options. e.g. \code{option(gh_appname = "YOUR_APP")} etc. This function will then automatically read those values. Otherwise specify them inline.
}
\examples{
\dontrun{
token <- github_auth()
}
}

