% Generated by roxygen2 (4.0.2): do not edit by hand
\name{github_stats}
\alias{github_stats}
\title{Generates a full list of GitHub stats and CRAN downloads from the RStudio mirror}
\usage{
github_stats(repo, verbose = TRUE)
}
\arguments{
\item{repo}{Name of a respository. Must include username or organization in format \code{username/repo}}

\item{verbose}{= TRUE Prints progress by default.}
}
\description{
Generates a full list of GitHub stats and CRAN downloads from the RStudio mirror
}
\examples{
\dontrun{
github_stats("ropensci/alm")
}
}

