% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_handle_error.R
\name{handle_error}
\alias{handle_error}
\title{handle_error}
\usage{
handle_error(e, script_name, alert_title = "Error")
}
\arguments{
\item{e}{Error passed from tryCatch.}

\item{script_name}{Name of the script in R/ to help locate error source.}

\item{alert_tite}{Title of popup alerting user to error. Defaults to "Error".}
}
\description{
Handle an error caught with tryCatch
This function prints the error message and context to the console, as well as sending an error log to Azure when `!golem::app_dev()`.
}
\examples{
if (interactive()) {
     tryCatch({
         stop("This is an error")
     }, error = function(e) {
         handle_error(e, "test_script.R", "test_app")
     })
}

#Console Output

#---------------------------------Script---------------------------------
#R/test_script.R
#---------------------------------Context--------------------------------
#tryCatch({
#    stop("This is an error")
#}, error = function(e) {
#    handle_error(e, "test_script.R", "test_app")
#})
#
#---------------------------------Message--------------------------------
#Error: This is an error


}
