% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{linebreak}
\alias{linebreak}
\title{linebreak adds a line break in the dashboard}
\usage{
linebreak()
}
\description{
\code{linebreak} generates a line break in the dashboard
}
\examples{
dashboard_open(data=iris) # other options: pathoutput=getwd() ...
dcpiechart(x=names(iris)[5])
linebreak()
dcbarchart(x=names(iris)[1] , gap=75)
linebreak()
dcpiechart(x=names(iris)[2])
linebreak()
dctable(index=names(iris)[5])
dashboard_launch(browse = FALSE) # Just generates files. Server is not launched
}

