% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countries_server.R
\name{mod_countries_server}
\alias{mod_countries_server}
\title{Country profile server-side function}
\usage{
mod_countries_server(id, con)
}
\arguments{
\item{id}{Internal parameter for Tabler.}

\item{con}{Shared SQL connection pool, created and closed by `app_server()`.}
}
\description{
A shiny Module.
}
