% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sectors_server.R
\name{mod_sectors_server}
\alias{mod_sectors_server}
\title{Sector profile server-side function}
\usage{
mod_sectors_server(id, con)
}
\arguments{
\item{id}{Internal parameter for Tabler.}

\item{con}{Shared SQL connection, created and closed by `app_server()`.}
}
\description{
A tabler Module.
}
