% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review-history.R
\name{rev_hours_airtable}
\alias{rev_hours_airtable}
\title{Generate historical data on logged review times in hours from airtable database.}
\usage{
rev_hours_airtable()
}
\value{
A `data.frame` with one row per issue and some key statistics.
}
\description{
Generate historical data on logged review times in hours from airtable database.
}
