% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{review_history}
\alias{review_history}
\title{Generate historical data on software reviews.}
\usage{
review_history(quiet = FALSE)
}
\arguments{
\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
(Invisibly) A `data.frame` with one row per issue and some key
statistics.
}
\description{
This is a reduced version of the `reviews_gh_data()` function, which returns
data only on dates of issue opening and closing, to be used to generate
historical patterns.
}
