% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{review_status}
\alias{review_status}
\title{Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.}
\usage{
review_status(open_only = TRUE, browse = TRUE, quiet = FALSE)
}
\arguments{
\item{open_only}{If `TRUE` (default), only extract information for currently
open issues.}

\item{browse}{If `TRUE` (default), open the results as a \pkg{DT} `datatable`
HTML page in default browser.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
A `data.frame` with one row per issue and some key statistics.
}
\description{
Generate a summary report for incoming Editor-in-Charge of current state of
all open software-review issues.
}
