% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{sanction_objective_labels}
\alias{sanction_objective_labels}
\title{Human-readable labels for GSDB sanction objectives}
\usage{
sanction_objective_labels
}
\description{
Human-readable labels for GSDB sanction objectives
}
