% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{sanctions_narrative}
\alias{sanctions_narrative}
\title{Build a Wikipedia-style paragraph describing sanctions applied against a country}
\usage{
sanctions_narrative(d, country_name, lookup)
}
\arguments{
\item{d}{data.table as returned by `fetch_sanctions()`}

\item{country_name}{Display name of the sanctioned country (article included, e.g. "the United Kingdom")}

\item{lookup}{Flat name -> code vector, as returned by `flatten_countries()`, used to name the senders}
}
\description{
Build a Wikipedia-style paragraph describing sanctions applied against a country
}
