% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradestatisticsshiny-package.R
\docType{package}
\name{tradestatisticsdashboard-package}
\alias{tradestatisticsdashboard}
\alias{tradestatisticsdashboard-package}
\title{tradestatisticsdashboard: Open Trade Statistics Dashboard}
\description{
Iteractive dashboard to explore Open Trade Statistics data, which consists in a curated version of the International Trade and Production Database for Estimation (ITPD-E) and the International Trade and Production Database for Simulation (ITPD-S).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/tradestatistics-dashboard}
  \item Report bugs at \url{https://github.com/pachadotdev/tradestatistics-dashboardissues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]
}

}
\keyword{internal}
