% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{write_output}
\alias{write_output}
\title{Write a tibble or data.frame to a CSV or Parquet file}
\usage{
write_output(x, path, format = "csv")
}
\arguments{
\item{x}{The data to write}

\item{path}{The file path to write to}

\item{format}{File format: 'csv' or 'parquet' (default: 'csv')}
}
\description{
Write a tibble or data.frame to a CSV or Parquet file
}
