% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{editor_status}
\alias{editor_status}
\title{Generate summary data of current and historical activities of all rOpenSci
editors}
\usage{
editor_status(quiet = FALSE, aggregation_period = "quarter")
}
\arguments{
\item{quiet}{If `FALSE`, display progress information on screen.}

\item{aggregation_period}{The time period for aggregation of timeline for
editorial loads. Must be one of "month", "quarter", or "semester".}
}
\value{
A list of four items:
\itemize{
\item{'status', a `data.frame` of current editor status, with one row per
editor and some key statistics.}
\item{'timeline_total', a `data.frame` representing the timelines for each
editor of total concurrent editorial load.}
\item{'timeline_new', a `data.frame` representing the timelines for each
editor of editorial load of new issues opened in each time period (default of
quarter-year.}
\item{'reviews', a `data.frame` with one row per issue and some key
statistics.} }
}
\description{
Generate summary data of current and historical activities of all rOpenSci
editors
}
