% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors-airtable.R
\name{editor_vacation_status}
\alias{editor_vacation_status}
\title{Get current vacation status of all rOpenSci editors.}
\usage{
editor_vacation_status(quiet = FALSE)
}
\value{
A `data.frame` with one row per editor and information on current
vacation status obtained from rOpenSci's airtable database.
}
\description{
Get current vacation status of all rOpenSci editors.
}
