\name{get_tweet}
\alias{get_tweet}
\title{Function for getting the latest tweet from the timeline of user \code{@si_petersson}}
\usage{
  get_tweet(x)
}
\arguments{
  \item{x}{The number of tweets to get}
}
\description{
  Get the latest tweet from @si_petersson's timeline. Please note that the authentication file (e.g. .Rdata file) needs to be added to AppArmor custom rules
}